python-service/requirements.txt
python-docx==1.1.0
flask==3.0.0
flask-cors==4.0.0




